pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "protpy"
version = "1.3.0"
description = "Python package for generating various biochemical, physicochemical and structural descriptors/features of protein sequences."
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "AJ McKenna", email = "amckenna41@qub.ac.uk" }
]
maintainers = [
    { name = "AJ McKenna", email = "amckenna41@qub.ac.uk" }
]
keywords = [
    "bioinformatics",
    "protein engineering",
    "python",
    "pypi",
    "machine learning",
    "aaindex",
    "protein descriptors",
    "physicochemical descriptors",
    "biochemical descriptors",
    "structural descriptors",
    "pySAR",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.9"
dependencies = [
    "aaindex>=1.2.0",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "varname>=0.15.1",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-flake8",
    "pytest-timeout",
    "biopython>=1.87",
]
docs = [
    "sphinx",
]

[project.urls]
Homepage = "https://github.com/amckenna41/protPy"
Repository = "https://github.com/amckenna41/protPy"
"Bug Tracker" = "https://github.com/amckenna41/protpy/issues"
Download = "https://github.com/amckenna41/protPy/archive/refs/heads/main.zip"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"protpy.data" = ["*.json"]

[tool.pytest.ini_options]
norecursedirs = ["docs/_build"]
